% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuboid_nd.R
\name{cuboid_nd}
\alias{cuboid_nd}
\title{Sample points from a n-dimensional rectangle prism (cuboid)}
\usage{
cuboid_nd(
  n_samples = 1000,
  n_dim = 3,
  dimnames = paste0("x", 1:n_dim),
  lengths = rep(1, n_dim),
  n_bins = 9,
  origin = rep(0, n_dim),
  return_coords_only = F,
  return_3D_radii = T,
  return_3D_angles = T,
  inside = T,
  rot_mat = diag(n_dim)
)
}
\arguments{
\item{rot_mat}{}
}
\description{
Sample points from a n-dimensional rectangle prism (cuboid)
}
