% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{ggbin}
\alias{ggbin}
\title{Title}
\usage{
ggbin(
  object,
  x = "x1",
  y = "x2",
  color = "name",
  facet_var = "x3_bin",
  ncol_legend = 2,
  legend.position = "bottom",
  theme = theme_void
)
}
\arguments{
\item{theme}{}
}
\description{
Title
}
