% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere_nd.R
\name{sphere_nd}
\alias{sphere_nd}
\title{Sample points from a n-dimensional sphere}
\usage{
sphere_nd(
  n_samples = 1000,
  n_dim = 3,
  dimnames = paste0("x", 1:n_dim),
  density_fun = function(r) 3 * r^2,
  radius = 1,
  scaling = rep(1, n_dim),
  origin = rep(0, n_dim),
  inside = T,
  n_bins = 9,
  return_3D_angles = T,
  return_3D_radii = T,
  return_coords_only = F
)
}
\arguments{
\item{return_coords_only}{}
}
\description{
Sample points from a n-dimensional sphere
}
