% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokamak.R
\name{tokamak}
\alias{tokamak}
\title{Title}
\usage{
tokamak(
  n_samples = 1000,
  R_plasma = 3,
  r_plasma = 0.05,
  R_magnet = 2,
  r_magnet = 0.1,
  RR_magnet = 3,
  magnet_freq = pi/8,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
