% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus_3d.R
\name{torus_3d}
\alias{torus_3d}
\title{Sample points from a 3-dimensional torus}
\usage{
torus_3d(
  n_samples = 1000,
  R = 3,
  r = 1,
  amplitude = 0,
  frequency = 2,
  n_dim = 3,
  dimnames = paste0("x", 1:n_dim),
  scaling = rep(1, n_dim),
  origin = rep(0, n_dim),
  inside = T,
  n_bins = 9,
  return_3D_angles = T,
  return_3D_radii = T,
  return_coords_only = F,
  rot_mat = diag(n_dim)
)
}
\arguments{
\item{return_3D_radii}{}
}
\description{
Sample points from a 3-dimensional torus
}
